pyproject.toml
[project]
name = "flight-prices"
version = "0.1.0"
description = "Monitor Azul flight prices VCP-MAD"
authors = [
    {name = "Matheus920"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "requests (>=2.33.0,<3.0.0)",
    "flights (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
